refactor(cli): extract hosted-registry installs into pack_remote.rs (LOC gate)
pack_cmd.rs crossed the 1500-line CI gate after the depth-1 dependency
resolution landed. The remote-install path (cmd_pack_install_remote,
install_declared_dependencies, cmd_pack_update) is a cohesive unit — it
moves to cli/pack_remote.rs unchanged; pack_cmd keeps the local store,
create, import/export and A2A commands.
